test: consolidate scattered review-round test files
The review round left many small single-purpose test files. Group them
without changing any test logic (verified: same total func count, full
suite green):

- internal_test.go (package convert, white-box): absorbs
  cache_bound_test.go + panic_sentinel_test.go
- review_misc_test.go (package convert_test): absorbs
  popitem_guard_test.go + timeval_test.go + conv_cleanup_test.go
- unwrap_test.go absorbs unwrap_fallback_test.go (same feature)
- checked_conv_test.go absorbs nil_elem_test.go (same theme)

Net -5 test files; no behavior change.
