+++ convert/internal_test.go
@@ -1,12 +1,17 @@
 package convert
 
 import (
+	"errors"
 	"fmt"
 	"reflect"
+	"strings"
 	"testing"
 
 	"go.starlark.net/starlark"
 )
+
+// Consolidated white-box (package convert) tests for internal helpers:
+// the bounded type-check cache and the conversion-boundary panic sentinel.
 
 // TestTypeCacheBounded verifies the type-check caches cannot grow without
 // bound. hashableGoValue mints a fresh reflect.ArrayOf(N, ...) type per
@@ -70,5 +75,38 @@ func TestBoundedCacheReturnsCachedAndFresh(t *testing.T) {
 	// second call hits the cache and must agree
 	if err := c.loadOrStore(mt, fmt.Errorf("should-not-be-used")); err != nil {
 		t.Fatalf("cached hit should return the stored nil error, got %v", err)
+	}
+}
+
+// TestToValuePanicSentinel verifies the conversion boundary's recover
+// produces a typed *PanicError carrying the recovered value and the stack
+// where the panic started, instead of a bare message that hides the
+// origin. The panic is provoked through reflect: reading an unexported
+// field yields a value whose Interface() panics.
+func TestToValuePanicSentinel(t *testing.T) {
+	type hidden struct {
+		secret string //nolint:unused // read via reflect to provoke the panic
+	}
+	rv := reflect.ValueOf(hidden{secret: "x"}).Field(0)
+	if rv.CanInterface() {
+		t.Fatal("test setup broken: expected an unexported field value")
+	}
+
+	v, err := toValue(rv, "")
+	if err == nil {
+		t.Fatalf("expected an error, got value %v", v)
+	}
+	var pe *PanicError
+	if !errors.As(err, &pe) {
+		t.Fatalf("expected *PanicError, got %T: %v", err, err)
+	}
+	if pe.Value == nil || len(pe.Stack) == 0 {
+		t.Fatalf("expected recovered value and stack, got %+v", pe)
+	}
+	if !strings.Contains(err.Error(), "panic recovered") {
+		t.Fatalf("expected historic message prefix, got %q", err.Error())
+	}
+	if !strings.Contains(string(pe.Stack), "toValue") {
+		t.Fatalf("expected stack to identify the conversion frame, got:\n%s", pe.Stack)
 	}
 }
